# laravelapi/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API Documentation",
        "description": "API documentation for your Laravel backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/wishlist": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get all wishlist items",
                "operationId": "7de9c05c67c97110c68860d9c0b814f9",
                "responses": {
                    "200": {
                        "description": "List of wishlist items"
                    }
                }
            },
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Add a new item to the wishlist",
                "operationId": "69cb7083b99e5e99c5ecbc02ff55e9af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "product_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wishlist item created"
                    }
                }
            }
        },
        "/api/wishlist/{id}": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Get wishlist item by ID",
                "operationId": "0462f5da607e16c2a0a859eaa24b0493",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist item details"
                    },
                    "404": {
                        "description": "Wishlist item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Delete a wishlist item",
                "operationId": "4baf028ba3d74d950330a8f0ec3acf55",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist item delete successful"
                    },
                    "404": {
                        "description": "Wishlist item not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Update an existing wishlist item",
                "operationId": "ea55781e3d6e9da495a29eacc9ffbfe7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wishlist item updated successfully"
                    },
                    "404": {
                        "description": "Wishlist item not found"
                    }
                }
            }
        }
    },
...
}